Updates   * Cleaned up sync task for repos   * Moved logic to run multiple syncs from the sync task into the sync service   * Moved original syncService logic to buildSync method   * Updated compose service to call the buildSync method
